Fixing db delete, Investigation still to be done on events table
