+++ models/event.js
@@ -1,16 +1,7 @@
 const mongoose = require('mongoose');
+const eventSchema = require('./eventSchema.js');
 
 var options = { discriminatorKey: 'type' };
-
-// Define the base Event schema
-const eventSchema = new mongoose.Schema({
-    _id: mongoose.Schema.Types.UUID,
-    type: String,
-    responseId: String,
-    isoTimestampSent: String,
-    isoTimestampReceived: String,
-}, 
-options);
 
 const Event = mongoose.model('Event', eventSchema);
