+++ models/response.js
@@ -1,0 +1,4 @@
+var mongoose = require('mongoose');  
+const responseSchema = require('./responseSchema.js');
+
+module.exports = mongoose.model('Response', responseSchema);
